requirements.txt
# -e .
aws-cdk-lib
constructs
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
cdk-ecr-deployment